requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-filter==24.2
pandas==2.2.2
psycopg2-binary==2.9.9
celery==5.4.0
redis==5.0.6
django-cacheops==7.0
python-dotenv==1.0.1
gunicorn==21.2.0
pytest==8.2.1
pytest-django==4.9.0
requests==2.32.3
